Ship (dolphin) no longer gets stuck in tiny gap
Two of the static lines formed a convex angle in the floor. The problem
is that if two lines don't cross each other, but only meet at a point,
then wedge-shaped objects can get stuck in them. This is a Chipmunk
Physics, general floating point precision, or fp accuracy problem.

Because the angle is convex, I couldn't make the lines cross one
another, or else it would not be possible to slide from one side to the
other. So I replaced those lines with a polygon, which I hope does not
share this problem. Doing this required that I allow polygons to be
static objects, and to change some other handling functions. Circles can
now be static too.