Moved over to CMake, CP library is now in tree
To make building easier, I've decided to use an in-tree copy of the
Chipmunk Physics library. I also stripped-down the CMakeLists.txt files
used by that project since I didn't include the demos here. (The extra
variables show up in graphical CMake configurators.)

I've also switched spacedolphin over to CMake (for better or for worse.)
Trying at first to use it to download the ChipmunkLatest.tgz file
automatically was a pain in the butt, but I'm not sure that it would
have been much better in a straight-up makefile if I had tried to
recover from download errors there too.

Finally, I've touched up the README.md file and restructured the source
tree a bit.