Set framerate to 55 FPS to avoid frame skipping
The framerate was originally set to 60 fps, which is the display rate of
my monitor. This would be fine, except that the render time for each
frame oscillates in a complex way such that at ~60 fps, some frames get
dropped on a continual basis.

This happens because the frame length usually goes short, long, short,
long, etc., so sometimes two frames are squeezed between two screen
refreshes, and only the last will be drawn. Then if the next frame time
is long, there may be no new frames copied into the frame buffer until
after the next refresh. Then the cycle may repeat, until the frequencies
change phase. The result is an occasional and persistent effective
framerate of about 1/2 of the maximum.

My solution to this problem is to keep the newer algorithm I chose for
setting the sleep time, but to set the framerate at 55 fps. I also tried
65 fps, but the 55 fps setting seemed smoother/more consistent. At this
rate, only one or two frames should be missed every 12 frames or so.

It would be good if I could find a way to detect the monitor's framerate
automatically, so that I could play nice with higher or lower framerates
on different machines.