Fixed frame time jitter. Now 60 fps works nicely.
The solution was to use a non-smoothed sleep time, set to the delta
between the actual frame time and the ideal. Also, it had to be right
after the render step and before the buffer flip step, so that the flip
would always occur at regular intervals, making for smooth graphics.

I also made a smoothed decrement to the sleep time, based on previous
deltas between requested and actual sleep time. This made for accurate
fps.