Improved turning control some more
Now there is a torque 'memory' that ramps down when you aren't holding
the left or right arrow, instead of cutting out completely. But no
torque is applied when a turning key isn't pressed. What this means is
that if you repeatedly press and release the left arrow after holding it
for a while, you will apply a mid-level of torque every time you press
the key.

Also, the torque curve is now based on the square root of the time the
key is held, not the 4th root, because of the lower initial torque that
makes finer control easier.