Changed black-hole gravity equation
The black hole equation was dividing the distance vector (between two
objects) by the distance squared. Since the vector was not scaled to a
unit vector, the gravity equation was proportional to one over the
distance, instead of one over the distance squared.

I fixed that, then created a repellant force close to the center of the
black hole. The difference is proportional to one over the distance
cubed.