Fix incorrect/absent fps on slower, 32 bit systems
On 32 bit systems, a 'long' is only 32 bits, so it can only represent a
few seconds' worth of nanoseconds. Since I was using that format
exclusively for frame durations, deltas and sums of time, the fps would
not display if the frame time was high. That is because it would
incorrectly think the sum of frame times was less than zero.

This has been partially fixed, though if individual frames take over two
seconds, this will still be an issue with the current setup. However, in
a situation like that, the game is unplayable, so it shouldn't be a
problem. Worst case, the fps will be wrong or not displayed.