Added new collision and color changing rules
now the rules are a little different. when the purple ship touches a
large object, it turns purple. when the purple large object touches a
small object, it turns blue. when the blue object touches the orange
ship, the small object loses its color. eventually, i plan on making
this a hp reduction for the ship, and adding health meters.

the same goes with the orange ship, etc.