Refactored the way color works. improved comments
instead of setting color on a per-shape basis, new shapes are marked
with properties that determines its color. the color is stored in a hash
table and found through a lookup function.

i improved the commenting here and there throught the code.

i also made some helpful changes to variable names.