improved color changing rules
now objects keep track of the last object they touched that changed
their color. this is so it can know whether it was recently re-touched
by the object, i.e. whether it is virtually maintaining contact.

if so, then the color will change to another color only if enough time
has lapsed, and is touched by an object of a different color.

when a small object touches a ship, it loses its color instanly, then
follows the above rules for regaining color.